src/use-cases/gcfaligner/faas/gcp/put/requirements.txt
google-cloud-storage == 2.1.0